refactor(react-app): start migrating to a table based view of streams
